Fix bug where replies and annotations were getting duplicated
The bug was caused by not checking the library id of the membership.
This means that a copy of each annotation/reply was given for each
membership in the database that matched the given user id. By checking
the library id of the membership against the current library, the bug
is fixed